requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
-e git+https://github.com/LCBRU/lbrc_flask.git@main#egg=lbrc_flask
    # via -r requirements.in
arrow==1.3.0
    # via lbrc-flask
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.1.3
    # via
    #   flask-security-too
    #   lbrc-flask
bleach==6.1.0
    # via flask-security-too
blinker==1.8.2
    # via
    #   flask
    #   flask-mail
    #   flask-principal
brotli==1.1.0
    # via fonttools
cachetools==5.3.3
    # via -r requirements.in
cffi==1.16.0
    # via weasyprint
chardet==5.2.0
    # via lbrc-flask
click==8.1.7
    # via flask
cssselect2==0.7.0
    # via weasyprint
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via
    #   flask-security-too
    #   lbrc-flask
et-xmlfile==1.1.0
    # via openpyxl
faker==25.2.0
    # via -r requirements.in
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-admin
    #   flask-api
    #   flask-login
    #   flask-mail
    #   flask-mailman
    #   flask-principal
    #   flask-security-too
    #   flask-sqlalchemy
    #   flask-weasyprint
    #   flask-wtf
    #   lbrc-flask
flask-admin==1.6.1
    # via lbrc-flask
flask-api==3.1
    # via lbrc-flask
flask-login==0.6.3
    # via flask-security-too
flask-mail==0.10.0
    # via lbrc-flask
flask-mailman==1.1.0
    # via flask-security-too
flask-principal==0.4.0
    # via flask-security-too
flask-security-too[common,fsqla]==5.4.3
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   flask-security-too
    #   lbrc-flask
flask-weasyprint==1.1.0
    # via lbrc-flask
flask-wtf==1.2.1
    # via flask-security-too
fonttools[woff]==4.52.1
    # via weasyprint
greenlet==3.0.3
    # via sqlalchemy
html5lib==1.1
    # via weasyprint
idna==3.7
    # via email-validator
importlib-metadata==7.1.0
    # via pygal
importlib-resources==6.4.0
    # via flask-security-too
itsdangerous==2.2.0
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.4
    # via flask
jsonschema==4.22.0
    # via lbrc-flask
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown==3.6
    # via lbrc-flask
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   flask-security-too
    #   jinja2
    #   werkzeug
    #   wtforms
mdurl==0.1.2
    # via markdown-it-py
openpyxl==3.1.2
    # via lbrc-flask
passlib==1.7.4
    # via flask-security-too
pillow==10.3.0
    # via weasyprint
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.4.0
    # via python-ldap
pycparser==2.22
    # via cffi
pydyf==0.10.0
    # via weasyprint
pygal==3.0.4
    # via lbrc-flask
pygments==2.18.0
    # via rich
pymysql==1.1.1
    # via -r requirements.in
pyphen==0.15.0
    # via weasyprint
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   faker
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   lbrc-flask
python-ldap==3.4.4
    # via lbrc-flask
pytz==2024.1
    # via lbrc-flask
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==13.7.1
    # via lbrc-flask
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   bleach
    #   html5lib
    #   python-dateutil
sqlalchemy==2.0.30
    # via
    #   -r requirements.in
    #   flask-security-too
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.2
    # via flask-security-too
tinycss2==1.3.0
    # via
    #   cssselect2
    #   weasyprint
types-python-dateutil==2.9.0.20240316
    # via arrow
typing-extensions==4.12.0
    # via sqlalchemy
weasyprint==62.1
    # via flask-weasyprint
webencodings==0.5.1
    # via
    #   bleach
    #   cssselect2
    #   html5lib
    #   tinycss2
werkzeug==3.0.3
    # via
    #   flask
    #   flask-login
wtforms==3.1.2
    # via
    #   flask-admin
    #   flask-security-too
    #   flask-wtf
xlrd==2.0.1
    # via lbrc-flask
zipp==3.18.2
    # via importlib-metadata
zopfli==0.2.3
    # via fonttools
